requirements.txt
Flask
Flask-MySQLdb
Flask-Bcrypt
Flask-Session
Flask-Mail
python-dotenv
Flask-WTF
waitress
pymysql